backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
marshmallow==3.25.1
google-generativeai>=0.8.0
fpdf2>=2.8.0
python-docx==1.1.0
